Update overview sidebar styles and add pdf analytics
