+++ src/components/tabs/tabs-styles.module.scss
@@ -13,7 +13,7 @@
     list-style-type: none;
     padding: 0;
     margin: 0;
-    border-radius: 2px;
+    border-radius: $card-border-radius;
     border: 1px solid $brand-color-main;
   }
