+++ src/containers/sidebars/local-scene-sidebar/overview-sidebar/styles.module.scss
@@ -1,0 +1,7 @@
+@import 'styles/settings';
+
+.actionButton {
+    margin-left: $site-gutter;
+    margin-bottom: $site-gutter;
+  }
+
